requirements-dev.txt
flake8==4.0.1
pytest-django==4.4.0
factory-boy==3.2.1
pytest-factoryboy==2.1.0

-r requirements.txt